fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@22436 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSParkingArea.cpp
@@ -154,7 +154,7 @@ MSParkingArea::addLotEntry(SUMOReal x, SUMOReal y, SUMOReal z,
 
 void
 MSParkingArea::enter(SUMOVehicle* what, SUMOReal beg, SUMOReal end) {
-    if (myLastFreeLot >= 1 && myLastFreeLot <= mySpaceOccupancies.size()) {
+    if (myLastFreeLot >= 1 && myLastFreeLot <= (int)mySpaceOccupancies.size()) {
         mySpaceOccupancies[myLastFreeLot].vehicle = what;
         myEndPositions[what] = std::pair<SUMOReal, SUMOReal>(beg, end);
         computeLastFreePos();
